fix readme indentations, feat mpca_lang_auto tests

+++ mpc.h
@@ -369,8 +369,8 @@ mpc_err_t *mpca_lang(int flags, const char *language, ...);
 mpc_err_t *mpca_lang_file(int flags, FILE *f, ...);
 mpc_err_t *mpca_lang_pipe(int flags, FILE *f, ...);
 mpc_err_t *mpca_lang_contents(int flags, const char *filename, ...);
-mpc_err_t	*mpca_lang_auto(int flags, const char *language, mpc_auto_parsers_t **parser_refs);
-mpc_err_t	*mpca_lang_auto_files(int flags, unsigned long amount, char **files, mpc_auto_parsers_t **parser_refs);
+mpc_err_t *mpca_lang_auto(int flags, const char *language, mpc_auto_parsers_t **parser_refs);
+mpc_err_t *mpca_lang_auto_files(int flags, int amount, char **files, mpc_auto_parsers_t **parser_refs);
 
 /*
 ** Misc
@@ -390,7 +390,7 @@ int mpc_test_fail(mpc_parser_t *p, const char *s, const void *d,
   int(*tester)(const void*, const void*),
   mpc_dtor_t destructor,
   void(*printer)(const void*));
-int mpc_auto_find_parser(char *name, mpc_auto_parsers_t	*autoparser, mpc_parser_t **parser_ref);
+int mpc_auto_find_parser(char *name, mpc_auto_parsers_t *autoparser, mpc_parser_t **parser_ref);
 void mpc_auto_delete(mpc_auto_parsers_t *autoparser);
 
 #ifdef __cplusplus
